backend/requirements.txt
anyio==4.12.1
asgiref==3.9.2
boto3==1.41.2
botocore==1.41.2
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
django-storages==1.14.6
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
requests==2.32.5
s3transfer==0.15.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
whitenoise==6.11.0
